adding bounds to training

+++ src/mitigated-fit/training.py
@@ -3,18 +3,19 @@ import argparse
 import json
 import os
 import time
+from functools import reduce
+from itertools import product
 
 import numpy as np
+from bp_utils import bound_pred
+from prepare_data import prepare_data
 from qibo import gates
 from qibo.backends import construct_backend
 from qibo.models.error_mitigation import calibration_matrix
-from qibo.noise import DepolarizingError, PauliError, NoiseModel
+from qibo.noise import DepolarizingError, NoiseModel, PauliError, ReadoutError
 from savedata_utils import get_training_type
 from uniplot import plot
-from prepare_data import prepare_data
 from vqregressor import vqregressor
-
-from bp_utils import bound_pred
 
 parser = argparse.ArgumentParser(description="Training the vqregressor")
 parser.add_argument("example")
@@ -38,25 +39,33 @@ layers = conf["nlayers"]
 ndata = conf["ndata"]
 
 # get string to identify the training type
-training_type = get_training_type(conf["mitigation"])
+training_type = get_training_type(conf["mitigation"], conf["noise"])
 
 # prepare data
 data, labels, scaler = prepare_data(conf["function"], show_sample=True)
 
 # noise model
 if conf["noise"]:
+    qm = 0.1
+    probabilities = [0.03,0.03,0.03]
+
+    paulis = list(product(["I", "X", "Y", "Z"], repeat=nqubits))[1:]
+    single_readout_matrix = np.array([[1-qm,qm],[qm,1-qm]])
+    readout_matrix = reduce(np.kron, [single_readout_matrix]*nqubits)
+    pauli_noise = PauliError(list(zip(paulis, probabilities)))
+    readout_noise = ReadoutError(readout_matrix)
+
     noise = NoiseModel()
-    error = PauliError(list(zip(["X", "Y", "Z"], [0.03, 0.03, 0.03])))
-    noise.add(error, gates.I)
-    #noise.add(DepolarizingError(lam=0.1), gates.RX)
+    noise.add(pauli_noise, gates.I)
+    noise.add(readout_noise, gates.M)
 else:
     noise = None
 
 if conf["qibolab"]:
     backend = construct_backend("qibolab", conf["platform"])
 else:
-    backend = construct_backend("qibojit", platform="numba")
-    #backend = construct_backend("numpy")
+    backend = construct_backend("qibojit", "numba")
+    backend.set_threads(1)
     
 readout = {}
 if conf["mitigation"]["readout"] is not None:
@@ -73,7 +82,7 @@ if conf["mitigation"]["readout"] is not None:
 
 mit_kwargs = {
     "ZNE": {"noise_levels": np.arange(5), "insertion_gate": "RX", "readout": readout},
-    "CDR": {"n_training_samples": 10, "readout": readout, "N_update": 10, "N_mean": 4},
+    "CDR": {"n_training_samples": 10, "readout": readout, "N_update": 20, "N_mean": 4},
     "vnCDR": {
         "n_training_samples": 10,
         "noise_levels": np.arange(3),
@@ -93,6 +102,7 @@ VQR = vqregressor(
     obs_hardware=conf["obs_hardware"],
     backend=backend,
     noise_model=noise,
+    bp_bound=conf["bp_bound"],
     mitigation=conf["mitigation"],
     mit_kwargs=mit_kwargs[conf["mitigation"]["method"]],
     scaler=scaler,
@@ -111,6 +121,7 @@ if conf["optimizer"] == "Adam":
         restart_from_epoch=conf["restart_from_epoch"],
         batchsize=conf["batchsize"],
         method="Adam",
+        J_treshold=2/conf["nshots"],
     )
 elif conf["optimizer"] == "CMA":
     VQR.cma_optimization()
@@ -128,9 +139,9 @@ print(f"Execution time required: ", (end - start))
 VQR.show_predictions(f"{args.example}/predictions_{conf['optimizer']}", save=True)
 np.save(f"{cache_dir}/best_params_{conf['optimizer']}_{training_type}", VQR.params)
 
-params = noise.errors[gates.I][0][1].options
-probs = [params[k][1] for k in range(3)]
-
-bounds = bound_pred(probs, layers, nqubits)
-print(bounds)
-np.save(f"{cache_dir}/pred_bound", np.array(bounds))
+if conf["noise"] and conf["bp_bound"] and os.path.exists(f"{cache_dir}/pred_bound") == False:
+    params = noise.errors[gates.I][0][1].options
+    probs = [params[k][1] for k in range(3)]
+    bit_flip = noise.errors[gates.M][0][1].options[0,-1]**(1/nqubits)
+    bounds = bound_pred(layers, nqubits, probs, bit_flip)
+    np.save(f"{cache_dir}/pred_bound", np.array(bounds))
